# openapi/generated/brew-note/model_create_bean.go
/*
brew-note api

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

API version: 0.0.1
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package openapi

import (
	"encoding/json"
)

// checks if the CreateBean type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &CreateBean{}

// CreateBean struct for CreateBean
type CreateBean struct {
	ProductionArea string `json:"productionArea"`
	PlantationName string `json:"plantationName"`
	Kind string `json:"kind"`
	RoastLevel string `json:"roastLevel"`
	Price NullableFloat64 `json:"price"`
}

// NewCreateBean instantiates a new CreateBean object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewCreateBean(productionArea string, plantationName string, kind string, roastLevel string, price NullableFloat64) *CreateBean {
	this := CreateBean{}
	this.ProductionArea = productionArea
	this.PlantationName = plantationName
	this.Kind = kind
	this.RoastLevel = roastLevel
	this.Price = price
	return &this
}

// NewCreateBeanWithDefaults instantiates a new CreateBean object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewCreateBeanWithDefaults() *CreateBean {
	this := CreateBean{}
	return &this
}

// GetProductionArea returns the ProductionArea field value
func (o *CreateBean) GetProductionArea() string {
	if o == nil {
		var ret string
		return ret
	}

	return o.ProductionArea
}

// GetProductionAreaOk returns a tuple with the ProductionArea field value
// and a boolean to check if the value has been set.
func (o *CreateBean) GetProductionAreaOk() (*string, bool) {
	if o == nil {
		return nil, false
	}
	return &o.ProductionArea, true
}

// SetProductionArea sets field value
func (o *CreateBean) SetProductionArea(v string) {
	o.ProductionArea = v
}

// GetPlantationName returns the PlantationName field value
func (o *CreateBean) GetPlantationName() string {
	if o == nil {
		var ret string
		return ret
	}

	return o.PlantationName
}

// GetPlantationNameOk returns a tuple with the PlantationName field value
// and a boolean to check if the value has been set.
func (o *CreateBean) GetPlantationNameOk() (*string, bool) {
	if o == nil {
		return nil, false
	}
	return &o.PlantationName, true
}

// SetPlantationName sets field value
func (o *CreateBean) SetPlantationName(v string) {
	o.PlantationName = v
}

// GetKind returns the Kind field value
func (o *CreateBean) GetKind() string {
	if o == nil {
		var ret string
		return ret
	}

	return o.Kind
}

// GetKindOk returns a tuple with the Kind field value
// and a boolean to check if the value has been set.
func (o *CreateBean) GetKindOk() (*string, bool) {
	if o == nil {
		return nil, false
	}
	return &o.Kind, true
}

// SetKind sets field value
func (o *CreateBean) SetKind(v string) {
	o.Kind = v
}

// GetRoastLevel returns the RoastLevel field value
func (o *CreateBean) GetRoastLevel() string {
	if o == nil {
		var ret string
		return ret
	}

	return o.RoastLevel
}

// GetRoastLevelOk returns a tuple with the RoastLevel field value
// and a boolean to check if the value has been set.
func (o *CreateBean) GetRoastLevelOk() (*string, bool) {
	if o == nil {
		return nil, false
	}
	return &o.RoastLevel, true
}

// SetRoastLevel sets field value
func (o *CreateBean) SetRoastLevel(v string) {
	o.RoastLevel = v
}

// GetPrice returns the Price field value
// If the value is explicit nil, the zero value for float64 will be returned
func (o *CreateBean) GetPrice() float64 {
	if o == nil || o.Price.Get() == nil {
		var ret float64
		return ret
	}

	return *o.Price.Get()
}

// GetPriceOk returns a tuple with the Price field value
// and a boolean to check if the value has been set.
// NOTE: If the value is an explicit nil, `nil, true` will be returned
func (o *CreateBean) GetPriceOk() (*float64, bool) {
	if o == nil {
		return nil, false
	}
	return o.Price.Get(), o.Price.IsSet()
}

// SetPrice sets field value
func (o *CreateBean) SetPrice(v float64) {
	o.Price.Set(&v)
}

func (o CreateBean) MarshalJSON() ([]byte, error) {
	toSerialize,err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o CreateBean) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	toSerialize["productionArea"] = o.ProductionArea
	toSerialize["plantationName"] = o.PlantationName
	toSerialize["kind"] = o.Kind
	toSerialize["roastLevel"] = o.RoastLevel
	toSerialize["price"] = o.Price.Get()
	return toSerialize, nil
}

type NullableCreateBean struct {
	value *CreateBean
	isSet bool
}

func (v NullableCreateBean) Get() *CreateBean {
	return v.value
}

func (v *NullableCreateBean) Set(val *CreateBean) {
	v.value = val
	v.isSet = true
}

func (v NullableCreateBean) IsSet() bool {
	return v.isSet
}

func (v *NullableCreateBean) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableCreateBean(val *CreateBean) *NullableCreateBean {
	return &NullableCreateBean{value: val, isSet: true}
}

func (v NullableCreateBean) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableCreateBean) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}